[Design] #29 - 구추(구리스의요청사항추가라는 뜻)
